% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_fit.R
\name{diagnostic_fit}
\alias{diagnostic_fit}
\title{Perform Diagnostics on CmdStan Fit}
\usage{
diagnostic_fit(fit, plot = TRUE, observed_y = NULL, delta = NULL)
}
\arguments{
\item{fit}{A CmdStanMCMC object returned from cmdstanr::sample()}

\item{plot}{Logical; whether to show diagnostic plots}

\item{observed_y}{Optional vector of observed values if predictive p-value is to be computed, the survey/survival length of each individual}

\item{delta}{Optional vector of censoring indicators (0 = censored, 1 = dead/failed) if predictive p-value is to be computed}
}
\value{
Invisible list of diagnostics
}
\description{
Perform Diagnostics on CmdStan Fit
}
