% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Animal Movement and Mortality Data}
\usage{
simulate_data(
  n_animals = 200,
  n_fixes = 1000,
  n_dead = 40,
  n_knots = 100,
  landscape_size = 1000
)
}
\arguments{
\item{n_animals}{Number of animals to simulate}

\item{n_fixes}{Maximum GPS fixes per animal}

\item{n_dead}{Number of dead animals (must be a squared number)}

\item{n_knots}{Number of spatial grid cells}

\item{landscape_size}{Size of landscape (in meters)}
}
\value{
A list containing all inputs for the Stan model
}
\description{
Simulate Animal Movement and Mortality Data
}
