% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_stan_data.R
\name{prepare_stan_data}
\alias{prepare_stan_data}
\title{Prepare Stan Data from Real Tracking Dataset}
\usage{
prepare_stan_data(
  df,
  include_hab_cov = TRUE,
  include_ind_cov = TRUE,
  hab_cov_names = c("prey_avail", "hunter"),
  ind_cov_names = c("age", "sex_m"),
  include_cell_effect = TRUE,
  grid_bounds = NULL,
  grid_res = NULL,
  cause_col = "delta",
  n_causes = NULL,
  max_time_gap = Inf
)
}
\arguments{
\item{df}{A data.frame containing GPS points with columns: x, y, animal_id, and any user-specified covariates}

\item{include_hab_cov}{Logical; whether to include habitat covariates}

\item{include_ind_cov}{Logical; whether to include individual-level covariates}

\item{hab_cov_names}{Character vector of habitat covariate column names}

\item{ind_cov_names}{Character vector of individual-level covariate column names}

\item{include_cell_effect}{Logical; whether to include spatial cell effect}

\item{grid_bounds}{Named list with xmin, xmax, ymin, ymax (only required if include_cell_effect = TRUE)}

\item{grid_res}{Grid resolution in meters (required if include_cell_effect = TRUE), please make sure that the grid bounds are divisible by the grid resolution}

\item{cause_col}{Character; name of the column in df that contains cause of death information}

\item{n_causes}{Integer; number of unique causes of death (if applicable)}

\item{max_time_gap}{Numeric; maximum time interval, in hours, allowed between consecutive GPS fixes. Intervals larger than this value are truncated to this value. The default is Inf, meaning that intervals are not truncated.}
}
\value{
A named list ready to be passed to a CmdStan model
}
\description{
Prepare Stan Data from Real Tracking Dataset
}
